Cc1 new session on load (#5)
* Updating studio project names for Snow 1, bow release

* pretty json .lofelt

* Clear session id on Studio projects
Loading these projects in the latest version of Lofelt Studio desktop will automatically create a new session and upload the audio and haptic data. This guarantees that every user opening this file will get a unique session and QR codes don't overlap between users

* Adding audio file relative path reference to .lofelt project
Legacy documents didn't write this path reference. It is needed for future versions of Studio to create a new session and upload the audio for auditioning